+++ src/modals/confirmation-modal/internal-operations-confirmation/internal-operations-confirmation.tsx
@@ -3,23 +3,23 @@ import React, { FC } from 'react';
 import { of } from 'rxjs';
 import { map, switchMap } from 'rxjs/operators';
 
+import { everstakeApi } from 'src/api.service';
 import { Disclaimer } from 'src/components/disclaimer/disclaimer';
+import { HeaderTitle } from 'src/components/header/header-title/header-title';
+import { useNavigationSetOptions } from 'src/components/header/use-navigation-set-options.hook';
+import { ApproveInternalOperationRequestActionPayloadInterface } from 'src/hooks/request-confirmation/approve-internal-operation-request-action-payload.interface';
+import { useRequestConfirmation } from 'src/hooks/request-confirmation/use-request-confirmation.hook';
+import { StacksEnum } from 'src/navigator/enums/stacks.enum';
+import { navigateAction } from 'src/store/root-state.actions';
+import { useSelectedRpcUrlSelector } from 'src/store/settings/settings-selectors';
+import { waitForOperationCompletionAction } from 'src/store/wallet/wallet-actions';
+import { useSelectedAccountSelector } from 'src/store/wallet/wallet-selectors';
+import { showSuccessToast } from 'src/toast/toast.utils';
+import { TEMPLE_WALLET_EVERSTAKE_LINK_ID } from 'src/utils/env.utils';
 import { isTruthy } from 'src/utils/is-truthy';
+import { RECOMMENDED_BAKER_ADDRESS } from 'src/utils/known-bakers';
+import { sendTransaction$ } from 'src/utils/wallet.utils';
 
-import { everstakeApi } from '../../../api.service';
-import { HeaderTitle } from '../../../components/header/header-title/header-title';
-import { useNavigationSetOptions } from '../../../components/header/use-navigation-set-options.hook';
-import { ApproveInternalOperationRequestActionPayloadInterface } from '../../../hooks/request-confirmation/approve-internal-operation-request-action-payload.interface';
-import { useRequestConfirmation } from '../../../hooks/request-confirmation/use-request-confirmation.hook';
-import { StacksEnum } from '../../../navigator/enums/stacks.enum';
-import { navigateAction } from '../../../store/root-state.actions';
-import { useSelectedRpcUrlSelector } from '../../../store/settings/settings-selectors';
-import { waitForOperationCompletionAction } from '../../../store/wallet/wallet-actions';
-import { useSelectedAccountSelector } from '../../../store/wallet/wallet-selectors';
-import { showSuccessToast } from '../../../toast/toast.utils';
-import { TEMPLE_WALLET_EVERSTAKE_LINK_ID } from '../../../utils/env.utils';
-import { sendTransaction$ } from '../../../utils/wallet.utils';
-import { RECOMMENDED_BAKER_ADDRESS } from '../../select-baker-modal/select-baker-modal';
 import { InternalOperationsConfirmationModalParams } from '../confirmation-modal.params';
 import { OperationsConfirmation } from '../operations-confirmation/operations-confirmation';
