+++ src/modals/select-baker-modal/select-baker-item/select-baker-item.tsx
@@ -15,12 +15,12 @@ import { formatSize } from 'src/styles/format-size';
 import { conditionalStyle } from 'src/utils/conditional-style';
 import { isDefined } from 'src/utils/is-defined';
 import { isTruthy } from 'src/utils/is-truthy';
+import { HELP_UKRAINE_BAKER_ADDRESS, RECOMMENDED_BAKER_ADDRESS } from 'src/utils/known-bakers';
 import { tzktUrl } from 'src/utils/linking.util';
 import { formatToPercentStr } from 'src/utils/number-format.utils';
 import { kFormatter } from 'src/utils/number.util';
 import { getTruncatedProps } from 'src/utils/style.util';
 
-import { HELP_UKRAINE_BAKER_ADDRESS, RECOMMENDED_BAKER_ADDRESS } from '../constants';
 import { useSelectBakerItemStyles } from './select-baker-item.styles';
 
 interface Props extends TestIdProps {
