+++ src/store/contact-book/contact-book-epics.ts
@@ -4,9 +4,10 @@ import { catchError, map, switchMap } from 'rxjs/operators';
 import { Action } from 'ts-action';
 import { ofType, toPayload } from 'ts-action-operators';
 
-import { loadTezosBalance$ } from '../../utils/token-balance.utils';
-import { withSelectedRpcUrl } from '../../utils/wallet.utils';
-import { RootState } from '../create-store';
+import { loadTezosBalance$ } from 'src/utils/token-balance.utils';
+import { withSelectedRpcUrl } from 'src/utils/wallet.utils';
+
+import type { RootState } from '../types';
 import { loadContactTezosBalance } from './contact-book-actions';
 
 const loadContactTezosBalanceEpic = (action$: Observable<Action>, state$: Observable<RootState>) =>
