+++ src/store/root-state.mock.ts
@@ -3,7 +3,6 @@ import { mockAdvertisingState } from './advertising/advertising-state.mock';
 import { mockBakingState } from './baking/baking-state.mock';
 import { mockBuyWithCreditCardState } from './buy-with-credit-card/state.mock';
 import { mockContactBookState } from './contact-book/contact-book-state.mock';
-import { RootState } from './create-store';
 import { mockCurrencyState } from './currency/currency-state.mock';
 import { mockDAppsState } from './d-apps/d-apps-state.mock';
 import { mockExolixState } from './exolix/exolix-state.mock';
@@ -14,6 +13,7 @@ import { mockSecurityState } from './security/security-state.mock';
 import { mockSettingsState } from './settings/settings-state.mock';
 import { mockSwapState } from './swap/swap-state.mock';
 import { mockTokensMetadataState } from './tokens-metadata/tokens-metadata-state.mock';
+import type { RootState } from './types';
 import { mockWalletState } from './wallet/wallet-state.mock';
 
 export const mockRootState: RootState = {
