+++ src/store/selector.ts
@@ -1,5 +1,5 @@
 import { TypedUseSelectorHook, useSelector as useSelectorRaw } from 'react-redux';
 
-import { RootState } from './create-store';
+import type { RootState } from './types';
 
 export const useSelector: TypedUseSelectorHook<RootState> = useSelectorRaw;
