+++ src/store/tokens-metadata/tokens-metadata-epics.ts
@@ -6,7 +6,7 @@ import { ofType, toPayload } from 'ts-action-operators';
 
 import { loadTokenMetadata$, loadTokensMetadata$, loadWhitelist$ } from '../../utils/token-metadata.utils';
 import { withSelectedRpcUrl } from '../../utils/wallet.utils';
-import { RootState } from '../create-store';
+import type { RootState } from '../types';
 import {
   addTokensMetadataAction,
   loadTokenMetadataActions,
